refactor(hadoop): consolidate credential conversion and validation
