+++ connectors/hadoop/src/test/java/io/unitycatalog/hadoop/internal/CredentialUtilTest.java
@@ -6,25 +6,100 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import io.unitycatalog.client.delta.model.DeltaCredentialOperation;
 import io.unitycatalog.client.delta.model.DeltaStorageCredential;
 import io.unitycatalog.client.delta.model.DeltaStorageCredentialConfig;
+import io.unitycatalog.client.model.AwsCredentials;
+import io.unitycatalog.client.model.AzureUserDelegationSAS;
+import io.unitycatalog.client.model.GcpOauthToken;
+import io.unitycatalog.client.model.TemporaryCredentials;
 import io.unitycatalog.hadoop.internal.auth.AwsCredential;
 import io.unitycatalog.hadoop.internal.auth.AzureCredential;
 import io.unitycatalog.hadoop.internal.auth.GcsCredential;
+import io.unitycatalog.hadoop.internal.auth.GenericCredential;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
-class DeltaStorageCredentialUtilTest {
+class CredentialUtilTest {
+  private static final long EXPIRATION = 123L;
+
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("validCredentials")
+  void convertsTemporaryCredentials(
+      String cloud, TemporaryCredentials input, GenericCredential expected) {
+    assertThat(CredentialUtil.toGenericCredential(input)).isEqualTo(expected);
+  }
+
+  private static Stream<Arguments> validCredentials() {
+    return Stream.of(
+        Arguments.of(
+            "AWS",
+            new TemporaryCredentials()
+                .awsTempCredentials(
+                    new AwsCredentials()
+                        .accessKeyId("access-key")
+                        .secretAccessKey("secret-key")
+                        .sessionToken("session-token"))
+                .expirationTime(EXPIRATION),
+            new AwsCredential("access-key", "secret-key", "session-token", EXPIRATION)),
+        Arguments.of(
+            "Azure",
+            new TemporaryCredentials()
+                .azureUserDelegationSas(new AzureUserDelegationSAS().sasToken("sas-token"))
+                .expirationTime(EXPIRATION),
+            new AzureCredential("sas-token", EXPIRATION)),
+        Arguments.of(
+            "GCS",
+            new TemporaryCredentials()
+                .gcpOauthToken(new GcpOauthToken().oauthToken("oauth-token"))
+                .expirationTime(EXPIRATION),
+            new GcsCredential("oauth-token", EXPIRATION)));
+  }
+
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("invalidCredentials")
+  void rejectsMissingRequiredField(
+      String cloud, TemporaryCredentials input, String expectedMessage) {
+    assertThatThrownBy(() -> CredentialUtil.toGenericCredential(input))
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessage(expectedMessage);
+  }
+
+  private static Stream<Arguments> invalidCredentials() {
+    return Stream.of(
+        Arguments.of(
+            "AWS",
+            new TemporaryCredentials()
+                .awsTempCredentials(
+                    new AwsCredentials()
+                        .secretAccessKey("secret-key")
+                        .sessionToken("session-token")),
+            "AWS access key is missing"),
+        Arguments.of(
+            "Azure",
+            new TemporaryCredentials()
+                .azureUserDelegationSas(new AzureUserDelegationSAS().sasToken("")),
+            "Azure SAS token is missing"),
+        Arguments.of(
+            "GCS",
+            new TemporaryCredentials().gcpOauthToken(new GcpOauthToken()),
+            "GCS OAuth token is missing"),
+        Arguments.of(
+            "no credential",
+            new TemporaryCredentials(),
+            "UC temporary credentials contained no cloud credential"));
+  }
 
   @Test
   void selectorRejectsMissingResponse() {
-    assertThatThrownBy(() -> DeltaStorageCredentialUtil.selectForLocation("s3://bucket/t", null))
+    assertThatThrownBy(() -> CredentialUtil.selectForLocation("s3://bucket/t", null))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("has no storage credentials");
     assertThatThrownBy(
-            () ->
-                DeltaStorageCredentialUtil.selectForLocation(
-                    "s3://bucket/t", Collections.emptyList()))
+            () -> CredentialUtil.selectForLocation("s3://bucket/t", Collections.emptyList()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("has no storage credentials");
   }
@@ -32,8 +107,7 @@ class DeltaStorageCredentialUtilTest {
   @Test
   void selectorRejectsSingleWithoutPrefixMatch() {
     DeltaStorageCredential only = credAt("s3://other-bucket");
-    assertThatThrownBy(
-            () -> DeltaStorageCredentialUtil.selectForLocation("s3://bucket/t", List.of(only)))
+    assertThatThrownBy(() -> CredentialUtil.selectForLocation("s3://bucket/t", List.of(only)))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("No UC Delta credential matched");
   }
@@ -42,8 +116,7 @@ class DeltaStorageCredentialUtilTest {
   void selectorRejectsSingleNull() {
     assertThatThrownBy(
             () ->
-                DeltaStorageCredentialUtil.selectForLocation(
-                    "s3://bucket/t", Collections.singletonList(null)))
+                CredentialUtil.selectForLocation("s3://bucket/t", Collections.singletonList(null)))
         .isInstanceOf(NullPointerException.class)
         .hasMessageContaining("contains null");
   }
@@ -54,33 +127,29 @@ class DeltaStorageCredentialUtilTest {
     DeltaStorageCredential table = credAt("s3://bucket/t");
     DeltaStorageCredential child = credAt("s3://bucket/t/child");
     assertThat(
-            DeltaStorageCredentialUtil.selectForLocation(
+            CredentialUtil.selectForLocation(
                 "s3://bucket/t/child/file", Arrays.asList(bucket, table, child)))
         .isSameAs(child);
   }
 
   @Test
   void selectorMatchesAtPathBoundary() {
-    assertThat(DeltaStorageCredentialUtil.prefixCovers("s3://bucket/t", "s3://bucket/t")).isTrue();
-    assertThat(DeltaStorageCredentialUtil.prefixCovers("s3://bucket/t/x", "s3://bucket/t"))
-        .isTrue();
-    assertThat(DeltaStorageCredentialUtil.prefixCovers("s3://bucket/t-other", "s3://bucket/t"))
-        .isFalse();
+    assertThat(CredentialUtil.prefixCovers("s3://bucket/t", "s3://bucket/t")).isTrue();
+    assertThat(CredentialUtil.prefixCovers("s3://bucket/t/x", "s3://bucket/t")).isTrue();
+    assertThat(CredentialUtil.prefixCovers("s3://bucket/t-other", "s3://bucket/t")).isFalse();
   }
 
   @Test
   void selectorNormalizesTrailingSlashes() {
-    assertThat(DeltaStorageCredentialUtil.prefixCovers("s3://bucket/t//", "s3://bucket/t"))
-        .isTrue();
-    assertThat(DeltaStorageCredentialUtil.prefixCovers("s3://bucket/t", "s3://bucket/t///"))
-        .isTrue();
+    assertThat(CredentialUtil.prefixCovers("s3://bucket/t//", "s3://bucket/t")).isTrue();
+    assertThat(CredentialUtil.prefixCovers("s3://bucket/t", "s3://bucket/t///")).isTrue();
   }
 
   @Test
   void selectorIgnoresNullAndPrefixlessInMultiResponse() {
     List<DeltaStorageCredential> creds =
         Arrays.asList(null, new DeltaStorageCredential(), credAt("s3://bucket/t"));
-    assertThat(DeltaStorageCredentialUtil.selectForLocation("s3://bucket/t", creds).getPrefix())
+    assertThat(CredentialUtil.selectForLocation("s3://bucket/t", creds).getPrefix())
         .isEqualTo("s3://bucket/t");
   }
 
@@ -88,7 +157,7 @@ class DeltaStorageCredentialUtilTest {
   void selectorThrowsWhenMultiResponseHasNoMatch() {
     List<DeltaStorageCredential> creds =
         Arrays.asList(credAt("s3://other"), credAt("s3://bucket/sibling"));
-    assertThatThrownBy(() -> DeltaStorageCredentialUtil.selectForLocation("s3://bucket/t", creds))
+    assertThatThrownBy(() -> CredentialUtil.selectForLocation("s3://bucket/t", creds))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("No UC Delta credential matched");
   }
@@ -105,7 +174,7 @@ class DeltaStorageCredentialUtilTest {
                     .s3AccessKeyId("ak")
                     .s3SecretAccessKey("sk")
                     .s3SessionToken("st"));
-    AwsCredential gc = (AwsCredential) DeltaStorageCredentialUtil.toGenericCredential(c);
+    AwsCredential gc = (AwsCredential) CredentialUtil.toGenericCredential(c);
     assertThat(gc.expirationTimeMillis()).isEqualTo(123L);
     assertThat(gc.accessKeyId()).isEqualTo("ak");
     assertThat(gc.secretAccessKey()).isEqualTo("sk");
@@ -119,7 +188,7 @@ class DeltaStorageCredentialUtilTest {
             .prefix("s3://bucket")
             .operation(DeltaCredentialOperation.READ)
             .config(new DeltaStorageCredentialConfig().s3AccessKeyId("ak").gcsOauthToken("gcs"));
-    assertThatThrownBy(() -> DeltaStorageCredentialUtil.toGenericCredential(c))
+    assertThatThrownBy(() -> CredentialUtil.toGenericCredential(c))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("must contain exactly one cloud credential config");
   }
@@ -128,7 +197,7 @@ class DeltaStorageCredentialUtilTest {
   void toGenericCredentialRejectsMissingConfig() {
     DeltaStorageCredential c =
         new DeltaStorageCredential().prefix("s3://bucket").operation(DeltaCredentialOperation.READ);
-    assertThatThrownBy(() -> DeltaStorageCredentialUtil.toGenericCredential(c))
+    assertThatThrownBy(() -> CredentialUtil.toGenericCredential(c))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("missing config");
   }
@@ -140,7 +209,7 @@ class DeltaStorageCredentialUtilTest {
             .prefix("abfss://container@account.dfs.core.windows.net/")
             .operation(DeltaCredentialOperation.READ_WRITE)
             .config(new DeltaStorageCredentialConfig().azureSasToken("sas-token"));
-    AzureCredential gc = (AzureCredential) DeltaStorageCredentialUtil.toGenericCredential(c);
+    AzureCredential gc = (AzureCredential) CredentialUtil.toGenericCredential(c);
     assertThat(gc.sasToken()).isEqualTo("sas-token");
     assertThat(gc.expirationTimeMillis()).isEqualTo(Long.MAX_VALUE);
   }
@@ -153,7 +222,7 @@ class DeltaStorageCredentialUtilTest {
             .operation(DeltaCredentialOperation.READ)
             .expirationTimeMs(456L)
             .config(new DeltaStorageCredentialConfig().gcsOauthToken("gcs-oauth-token"));
-    GcsCredential gc = (GcsCredential) DeltaStorageCredentialUtil.toGenericCredential(c);
+    GcsCredential gc = (GcsCredential) CredentialUtil.toGenericCredential(c);
     assertThat(gc.oauthToken()).isEqualTo("gcs-oauth-token");
     assertThat(gc.expirationTimeMillis()).isEqualTo(456L);
   }
@@ -165,7 +234,7 @@ class DeltaStorageCredentialUtilTest {
             .prefix("s3://bucket")
             .operation(DeltaCredentialOperation.READ)
             .config(new DeltaStorageCredentialConfig().s3SessionToken("st"));
-    assertThatThrownBy(() -> DeltaStorageCredentialUtil.toGenericCredential(c))
+    assertThatThrownBy(() -> CredentialUtil.toGenericCredential(c))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("AWS access key is missing");
   }
